deploy/requirements_simple.txt
freqtrade[plot]
ccxt>=4.0.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
ft-pandas-ta>=0.3.14b0
python-telegram-bot>=20.0
aiohttp>=3.8.0
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.0
tables>=3.7.0
blosc>=1.10.0
pyarrow>=10.0.0
